Refactor sales ETL and schema for FactSales redesign
Redesigned FactSales and staging.sales tables to use composite keys (SaleID, OrderId, authorId) and updated related stored procedures and ETL packages. Enhanced GetSalesChangeByRowVersion to join additional tables and return more detailed sales data. Updated DW_MergeFactSales logic to match new keys, expanded DimDate data range, and added new FactSales ETL package. Changed sales.qty column type from smallint to int for consistency.

+++ Pubs/Schema/Tables/sales.sql
@@ -2,7 +2,7 @@
 	[stor_id] [char](4) NOT NULL,
 	[ord_num] [varchar](20) NOT NULL,
 	[ord_date] [datetime] NOT NULL,
-	[qty] [smallint] NOT NULL,
+	[qty] [int] NOT NULL,
 	[payterms] [varchar](12) NOT NULL,
 	[title_id] [varchar] (6) NOT NULL,
 	[RowVersion] [timestamp] NOT NULL,
